Moved some files around, and moved some common things into general files to be imported
